+++ screens/booking/MyBooking.tsx
@@ -2,7 +2,7 @@ import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import React from "react";
 import { View, StyleSheet } from "react-native";
 
-import MyBookingTab from "../../navigation/MyBookingTab";
+import MyBookingTab from "../../components/booking/MyBookingTab";
 import { RootParamList } from "../../types";
 
 export type MyBookingProps = NativeStackScreenProps<RootParamList, "MyBooking">;
